+++ release.config.js
@@ -13,26 +13,23 @@ module.exports = {
     branches: [
         {
             name: "master",
-            level: "minor",
+            prerelease: true,
+            level: "none",
+            channel: "next",
             devDependencies: ["@zowe/cli", "@zowe/imperative", "@zowe/cli-test-utils"]
         },
         {
-            name: "zowe-v1-lts",
+            name: "zowe-v?-lts",
             level: "patch",
             devDependencies: ["@zowe/cli", "@zowe/imperative"]
-        },
-        {
-            name: "next",
-            prerelease: true,
-            level: "next",
-            devDependencies: ["@zowe/cli", "@zowe/imperative", "@zowe/cli-test-utils"]
         }
     ],
     plugins: [
         "@octorelease/changelog",
         ["@octorelease/npm", {
             aliasTags: {
-                latest: ["zowe-v2-lts"]
+                latest: ["zowe-v2-lts"],
+                next: ["zowe-v3-lts"]
             },
             pruneShrinkwrap: true
         }],